Disabled consecutive database submissions

+++ app/api/chatbot.py
@@ -1,11 +1,17 @@
 
-from flask import request, abort, jsonify
+from flask import request, abort, jsonify, session
 from flask_login import current_user
 
 from . import api
 from .. import db
 from ..utils import text_generation
 from ..models import ChatbotMessage
+
+
+@api.before_request
+def before_chatbot_request():
+    if "is_chatbot_ready" not in session:
+        session["is_chatbot_ready"] = True
 
 
 @api.post("/chatbot/user-msg/create")
@@ -18,6 +24,9 @@ def create_user_message():
     value = data.get("chatbot-user-msg-create").strip()
     if current_user.is_authenticated:
         try:
+            if not session["is_chatbot_ready"]:
+                abort(400)
+            session["is_chatbot_ready"] = False
             item = ChatbotMessage(user=current_user, value=value, type="user")
         except ValueError:
             abort(400)
@@ -38,13 +47,16 @@ def edit_user_message(phase):
 
     # might try implementing a dict[str, method]
     if phase == "user":
+        if not session["is_chatbot_ready"]:
+            abort(400)
+        session["is_chatbot_ready"] = False
         if not "chatbot-user-msg-update" in data and not "id" in data:
             abort(400)
         user_msg_id = data.get("id")
         user_msg_value = data.get("chatbot-user-msg-update").strip()
         if current_user.is_authenticated:
             try:
-                item = db.session.execute(db.select(ChatbotMessage).filter_by(user=current_user, id=user_msg_id)).scalar_one()
+                item = db.session.execute(db.select(ChatbotMessage).where(db.and_(ChatbotMessage.user == current_user, ChatbotMessage.id == user_msg_id))).scalar_one()
                 setattr(item, "value", user_msg_value)
                 db.session.execute(db.delete(ChatbotMessage).where(db.and_(ChatbotMessage.user == current_user, ChatbotMessage.id >= item.id + 2)))
             except ValueError:
@@ -62,12 +74,13 @@ def edit_user_message(phase):
         print(server_msg_id, user_msg_last_value, server_msg_next_value)
         if current_user.is_authenticated:
             try:
-                item = db.session.execute(db.select(ChatbotMessage).filter_by(user=current_user, id=server_msg_id)).scalar_one()
+                item = db.session.execute(db.select(ChatbotMessage).where(db.and_(ChatbotMessage.user == current_user, ChatbotMessage.id == server_msg_id))).scalar_one()
                 setattr(item, "value", server_msg_next_value)
             except ValueError:
                 abort(400)
             else:
                 db.session.commit()
+                session["is_chatbot_ready"] = True
         return jsonify({
             "value": server_msg_next_value,
         })
@@ -84,9 +97,12 @@ def delete_user_message():
         abort(400)
     id = data.get("id")
     if current_user.is_authenticated:
-        item = db.session.execute(db.select(ChatbotMessage).filter_by(user=current_user, id=id)).scalar_one()
-        db.session.execute(db.delete(ChatbotMessage).where(db.and_(ChatbotMessage.user == current_user, ChatbotMessage.id >= item.id)))
+        if not session["is_chatbot_ready"]:
+            abort(400)
+        session["is_chatbot_ready"] = False
+        db.session.execute(db.delete(ChatbotMessage).where(db.and_(ChatbotMessage.user == current_user, ChatbotMessage.id >= id)))
         db.session.commit()
+        session["is_chatbot_ready"] = True
     return jsonify({})
 
 
@@ -107,6 +123,7 @@ def create_server_message():
         else:
             db.session.add(item)
             db.session.commit()
+            session["is_chatbot_ready"] = True
     return jsonify({
         "id": item.id,
         "value": server_msg_value,
@@ -115,6 +132,9 @@ def create_server_message():
 
 @api.post("/chatbot/server-msg/update")
 def update_server_message():
+    if not session["is_chatbot_ready"]:
+        abort(400)
+    session["is_chatbot_ready"] = False
     if not request.json:
         abort(400)
     data = request.get_json()
@@ -123,7 +143,7 @@ def update_server_message():
     # also delete all messages after
     id = data.get("id")
     if current_user.is_authenticated:
-        item = db.session.execute(db.select(ChatbotMessage).filter_by(user=current_user, id=id)).scalar_one()
+        item = db.session.execute(db.select(ChatbotMessage).where(db.and_(ChatbotMessage.user == current_user, ChatbotMessage.id == id))).scalar_one()
         db.session.execute(db.delete(ChatbotMessage).where(db.and_(ChatbotMessage.user == current_user, ChatbotMessage.id >= item.id + 1)))
 
     user_msg_last_value = data.get("chatbot-user-msg-last")
@@ -135,6 +155,7 @@ def update_server_message():
             abort(400)
         else:
             db.session.commit()
+            session["is_chatbot_ready"] = True
     return jsonify({
         "value": server_msg_next_value,
     })
