+++ app/api/todolist.py
@@ -1,10 +1,16 @@
 
-from flask import request, abort, jsonify
+from flask import request, abort, jsonify, session
 from flask_login import current_user
 
 from . import api
 from .. import db
 from ..models import Todolist
+
+
+@api.before_request
+def before_todolist_request():
+    if "is_todolist_ready" not in session:
+        session["is_todolist_ready"] = True
 
 
 @api.post("/todolist/create")
@@ -16,6 +22,9 @@ def create_todolist_item():
         abort(400)
     value = data.get("todolist-create").strip()
     if current_user.is_authenticated:
+        if not session["is_todolist_ready"]:
+            abort(418)
+        session["is_todolist_ready"] = False
         try:
             item = Todolist(user=current_user, value=value)
         except ValueError:
@@ -23,6 +32,7 @@ def create_todolist_item():
         else:
             db.session.add(item)
             db.session.commit()
+            session["is_todolist_ready"] = True
     return jsonify({
         "id": item.id if current_user.is_authenticated else 0,
         "value": value,
@@ -39,13 +49,17 @@ def update_todolist_item():
     id = data.get("id")
     value = data.get("todolist-update").strip()
     if current_user.is_authenticated:
+        if not session["is_todolist_ready"]:
+            abort(418)
+        session["is_todolist_ready"] = False
         try:
-            item = db.session.execute(db.select(Todolist).filter_by(user=current_user, id=id)).scalar_one()
+            item = db.session.execute(db.select(Todolist).where(db.and_(Todolist.user == current_user, Todolist.id == id))).scalar_one()
             setattr(item, "value", value)
         except ValueError:
             return abort(400)
         else:
             db.session.commit()
+            session["is_todolist_ready"] = True
     return jsonify({})
 
 
@@ -57,6 +71,18 @@ def delete_todolist_item():
     if not "id" in data:
         abort(400)
     if current_user.is_authenticated:
+        if not session["is_todolist_ready"]:
+            abort(418)
+        session["is_todolist_ready"] = False
         db.session.execute(db.delete(Todolist).where(db.and_(Todolist.user == current_user, Todolist.id == data.get("id"))))
         db.session.commit()
+        session["is_todolist_ready"] = True
     return jsonify({})
+
+
+"""
+focus:
+none: parent(bg-sub-alt, text-sub)
+parent focus: bg-sub, text-text
+child focus: parent(bg-sub-alt, text-main), child same as parent
+"""
